fix: correct link to the chromium versions file
The `fetchCompatibilityVersions` function in the puppeteer_utils.js file retrieves a markdown file from the puppeteer github repo, and uses this file to determine which versions of Chromium are compatible with which versions of Puppeteer.

This information was previously stored in the file `docs/api.md`, but was recently moved to `docs/chromium-support.md`.  This commit gets the file from its new location.

In addition to changing the location of the file, there were two other changes to the contents of the `chromium-support.md` file.  

The first change was that the file no longer had the same text at the start and end of the compatibility list.  Previously, this code was looking for `'Releases per Chromium Version:'` and `'* [All releases]'` to indicate the beginning and end of the compatibility list.  The file now contains some helpful markdown comments that can be used to indicate the start and end of this list, specifically `'<!-- version-start -->'` and `'<!-- version-end -->'`.  This commit updates the  puppeteer_utils.js file to now look for these comments.

The second change was that the indentation on the list was slightly deeper, and this broke the code that filtered for lines that started with `* Chromium`.  This commit updates the puppeteer_utils.js file to filter based on the deeper `- Chromium` text instead.