feat: add script to ensure playwright/puppeteer version in package.json is in sync with docker tag (#177)
This will be used in the pw/pup templates as part of their dockerfile,
allowing to pin the versions in the templates.

The script fails only if it finds pw/pup in the package.json as well as
base docker image tag, otherwise exits with 0.

We might also need a script to update those pinned versions in the
templates.